refs #1790 allow File and ReadOnlyFile objects to be used in binary modules
