+++ include/qore/intern/QC_File.h
@@ -4,7 +4,7 @@
 
   Qore Programming Language
 
-  Copyright (C) 2003 - 2015 David Nichols
+  Copyright (C) 2003 - 2017 Qore Technologies, s.r.o.
 
   Permission is hereby granted, free of charge, to any person obtaining a
   copy of this software and associated documentation files (the "Software"),
@@ -33,39 +33,11 @@
 
 #define _QORE_CLASS_FILE_H
 
-DLLEXPORT extern qore_classid_t CID_FILE;
-DLLEXPORT extern QoreClass *QC_FILE;
-DLLEXPORT extern QoreClass *QC_READONLYFILE;
-
 DLLLOCAL QoreClass *initFileClass(QoreNamespace &qorens);
 
 #include <qore/QoreFile.h>
-#include <qore/AbstractPrivateData.h>
 
-class File : public AbstractPrivateData, public QoreFile {
-   protected:
-      DLLLOCAL virtual ~File() {}
-
-   public:
-      DLLLOCAL File(const QoreEncoding *cs) : QoreFile(cs) {}
-
-      DLLLOCAL virtual void deref(ExceptionSink *xsink) {
-         if (ROdereference()) {
-            cleanup(xsink);
-            delete this;
-         }
-      }
-
-      DLLLOCAL virtual void deref() {
-         if (ROdereference()) {
-            ExceptionSink xsink;
-            cleanup(&xsink);
-            delete this;
-         }
-      }
-};
-
-inline int check_terminal_io(QoreObject* self, const char* m, ExceptionSink* xsink) {
+DLLLOCAL inline int check_terminal_io(QoreObject* self, const char* m, ExceptionSink* xsink) {
    // check for no-terminal-io at runtime with system objecs
    if (self->isSystemObject() && runtime_check_parse_option(PO_NO_TERMINAL_IO)) {
       xsink->raiseException("ILLEGAL-EXPRESSION", "%s() cannot be called with a system constant object when 'no-terminal-io' is set", m);
